# artifacts/specs/openapi.yaml
openapi: 3.0.3
info:
  title: Talkam Liberia API
  version: "1.0.0"
  description: |
    REST API for Talkam Liberia mobile/web apps. Supports JWT + anonymous ephemeral tokens.
servers:
  - url: https://api.talkamliberia.org/v1
security:
  - bearerAuth: []
paths:
...
  /reports/create:
    post:
      summary: Create a report
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReportCreateRequest'
      responses:
        '201':
          description: Report created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportResponse'
        '409': { description: Duplicate detected }
  /reports/{id}:
    get:
      summary: Fetch report details
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200': { content: { application/json: { schema: { $ref: '#/components/schemas/ReportResponse' } } } }
        '404': { description: Not found }
  /reports/{id}/verify:
    post:
      summary: Submit verification action
    
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerificationRequest'
      responses:
        '200': { description: Verification recorded }
        '403': { description: Not authorized }
  /reports/{id}/comment:
    post:
      summary: Add comment or request info
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommentRequest'
      responses:
        '201': { description: Comment created }
  /reports/search:
    get:
      summary: Search reports
      parameters:
        - in: query
          name: county
          schema: { type: string }
        - in: query
          name: category
          schema: { type: string }
        - in: query
          name: status
          schema: { type: string, enum: [submitted, triaged, verified, resolved] }
        - in: query
          name: text
          schema: { type: string }
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                type: object
                properties:
                  results:
                    type: array
                    items:
                      $ref: '#/components/schemas/ReportSummary'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ReportCreateRequest:
      type: object
      required: [category, summary, location]
      properties:
        category: { type: string }
        severity: { type: string, enum: [low, medium, high, critical] }
        anonymous: { type: boolean }
        summary: { type: string }
        details: { type: string }
        media: { type: array, items: { $ref: '#/components/schemas/MediaRef' } }
        location: { $ref: '#/components/schemas/Location' }
        witness_count: { type: integer }
        offline_reference: { type: string }
        sms_reference: { type: string }
    ReportResponse:
      type: object
      properties:
        id: { type: string }
        status: { type: string }
        created_at: { type: string, format: date-time }
        summary: { type: string }
        details: { type: string }
        category: { type: string }
        severity: { type: string }
        location: { $ref: '#/components/schemas/Location' }
        verification_score: { type: number }
        media: { type: array, items: { $ref: '#/components/schemas/MediaRef' } }
        timeline: { type: array, items: { $ref: '#/components/schemas/TimelineEvent' } }
    ReportSummary:
      type: object
      properties:
        id: { type: string }
        summary: { type: string }
        category: { type: string }
        county: { type: string }
        severity: { type: string }
        status: { type: string }
    MediaRef:
      type: object
      properties:
        key: { type: string }
        type: { type: string, enum: [photo, video, audio] }
        checksum: { type: string }
        blur_faces: { type: boolean }
    Location:
      type: object
      properties:
        latitude: { type: number }
        longitude: { type: number }
        county: { type: string }
        district: { type: string }
        description: { type: string }
    TimelineEvent:
      type: object
      properties:
        at: { type: string, format: date-time }
        action: { type: string }
        actor: { type: string }
        notes: { type: string }
    VerificationRequest:
      type: object
      properties:
        action: { type: string, enum: [confirm, request_info, reject] }
        comment: { type: string }
        witness_count: { type: integer }
        attachments: { type: array, items: { $ref: '#/components/schemas/MediaRef' } }
    CommentRequest:
      type: object
      properties:
        body: { type: string }
        visibility: { type: string, enum: [reporter, ngo, admin, public] }
...